backend/agent/pyproject.toml
[project]
name = "logfire-docs-chatbot"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "bs4>=0.0.2",
    "email-validator>=2.2.0",
    "fastapi>=0.117.1",
    "lancedb>=0.25.0",
    "langchain-text-splitters>=0.3.11",
    "logfire[fastapi]>=4.9.0",
    "markdown2>=2.5.4",
    "pip>=25.2",
    "psycopg2-binary>=2.9.9",
    "pydantic-ai-slim[anthropic,google,openai]>=1.22.0",
    "pydantic-settings>=2.6.0",
    "pyright>=1.1.405",
    "python-dotenv>=1.0.0",
    "python-frontmatter>=1.1.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "sse-starlette>=3.0.2",
    "starlette>=0.48.0",
    "uvicorn>=0.37.0",
    "watchfiles>=1.1.0",
]

[tool.setuptools.packages.find]
include = ["chatbot*"]
exclude = ["testing*", "alembic*"]

[tool.ruff.format]
# don't format python in docstrings, pytest-examples takes care of it
docstring-code-format = false
quote-style = "single"
